Update model and implement user store
The UserStore is a custom version of ASP.NET Identity's UserManager
class. It implements a lot of user validation logic for free. I used
BCrypt for password hashing because it allows you to control how many
encryption rounds are performed, which is a time saver for generating
password hashes for tests. Normally I don't like putting a bunch of
changes into a single commit like this but it was quite a process
getting ASP.NET Identity to work with my models and migrate correctly.